Split Extensions/LanguageExtensions.cs into Core.Windows and an app-side RTL facade
Batch 3d. XmlLanguage comes from PresentationCore (WPF), so it cannot
cross into Text-Grab.Core.Windows, and reconnaissance confirmed the RTL
path is reachable in production from GrabFrame, EditTextWindow and
OcrUtilities - not a dead using.

- Text-Grab.Core.Windows/Extensions/LanguageExtensions.cs (namespace
  Text_Grab, class name kept exactly) - IsSpaceJoining (both the
  Windows.Globalization.Language and ILanguage overloads), IsLatinBased,
  AsLanguage, AsILanguage. Keeping the original type name means every
  existing call site in the app (OcrUtilities, GrabFrame, HistoryInfo,
  LanguageService, PdfDocumentRenderer, ...) resolves unchanged through
  the existing project reference - no call-site edits were needed at
  all, confirmed by a clean Tests.csproj build.
- Text-Grab/Extensions/LanguageRtlExtensions.cs (new file, same
  Text_Grab namespace, new type name) - IsRightToLeft(this Language)
  (the XmlLanguage lookup) and IsRightToLeft(this ILanguage) (the
  GlobalLang branch calls the Language overload, so the two stay
  together). Named distinctly rather than reusing LanguageExtensions:
  since both classes end up in the same Text_Grab namespace once the
  app references Core.Windows, a same-named extension method in both
  would have been a CS0121 ambiguity at every call site instead of one
  clean split.

No package references or settings touched. No files deferred from this
batch's list.

Gate: dotnet build Tests/Tests.csproj -c Debug -p:Platform=x64 - clean.
Tests.Core 514, Tests.Core.Windows 3, Tests 993 passed/7 skipped -
identical to baseline, total 1510 conserved.
