+++ Text-Grab.Core.Windows/Extensions/LanguageExtensions.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Globalization;
-using System.Windows.Markup;
 using Text_Grab.Interfaces;
 using Text_Grab.Models;
 using Windows.Globalization;
@@ -18,13 +16,6 @@ public static class LanguageExtensions
         return true;
     }
 
-    public static bool IsRightToLeft(this Language language)
-    {
-        XmlLanguage lang = XmlLanguage.GetLanguage(language.LanguageTag);
-        CultureInfo culture = lang.GetEquivalentCulture();
-        return culture.TextInfo.IsRightToLeft;
-    }
-
     public static bool IsSpaceJoining(this ILanguage selectedLanguage)
     {
         if (selectedLanguage.LanguageTag.StartsWith("zh", StringComparison.InvariantCultureIgnoreCase))
@@ -32,15 +23,6 @@ public static class LanguageExtensions
         else if (selectedLanguage.LanguageTag.Equals("ja", StringComparison.InvariantCultureIgnoreCase))
             return false;
         return true;
-    }
-
-    public static bool IsRightToLeft(this ILanguage selectedLanguage)
-    {
-        if (selectedLanguage is GlobalLang language)
-            return language.OriginalLanguage.IsRightToLeft();
-
-        // For other language types, use the LayoutDirection property
-        return selectedLanguage.LayoutDirection == LanguageLayoutDirection.Rtl;
     }
 
     public static bool IsLatinBased(this ILanguage selectedLanguage)
